Распределяет код по модулям
